% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_traces}
\alias{import_traces}
\title{Import instrument data files}
\usage{
import_traces(path, cutoff = 5, sheet = 1)
}
\arguments{
\item{path}{a character string path to the instrument export Excel file containing the raw data}

\item{cutoff}{a numeric value specifiying how many cycles to ignore
(if the assay requries the instrument lid to be opened for reagent addition after temperature equilibration);
defaults to *5*}

\item{sheet}{a numeric value representing the Excel workbook sheet which contains the raw data;
defaults to the first sheet, i.e. *1*}
}
\value{
a [tibble][tibble::tibble-package] in "long" form containing the variables *cycle*, *well*, and *intensity*
}
\description{
\code{import_traces} imports Excel data files exported from Bio-Rad real-time fluorescnce thermal cyclers.
}
