% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{view_model}
\alias{view_model}
\title{Generate a plot from real-time fluorescence data model parameters}
\usage{
view_model(
  data,
  control = NULL,
  color = well,
  size = NULL,
  shape = NULL,
  palette = "Set3",
  ncolors = 12,
  interactive = TRUE
)
}
\arguments{
\item{data}{a dataframe containing the imported real-time fluorescence data, variables *cycle*, *well*, and *intensity*}

\item{control}{a character string indicating the well to mark as the baseline control (e.g. *"A1"* or *"H12"*);
defaults to NULL (i.e. no control point marked on the plot)}

\item{color}{grouping variable to map to `geom_point` aesthetic *color*;
defaults to the *well* variable}

\item{size}{grouping variable to map to `geom_point` aesthetic *size*}

\item{shape}{grouping variable to map to `geom_point` aesthetic *shape*}

\item{palette}{character string indicating which ColorBrewer palette to use;
defaults to "Set3"}

\item{ncolors}{numeric value indicating how many colors from the specified palette to use to generate a spectrum of colors for the *color* grouping variable;
defaults to 12 for palette "Set3"}

\item{interactive}{boolean parameter to display a static or interactive plot using `plotly`;
defaults to TRUE}
}
\value{
a [tibble][tibble::tibble-package] containing the calculated variables *rate* and *endpoint*
}
\description{
\code{view_model} generates a plot of the variables derived from the model fit to real-time fluorescence data. Endpoint fluorescence is plotted against rate (rise time).
}
