% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{view_traces}
\alias{view_traces}
\title{View real-time fluorescence traces}
\usage{
view_traces(data, color = well, model = FALSE, interactive = TRUE)
}
\arguments{
\item{data}{a dataframe containing the imported real-time fluorescence data, variables *cycle*, *well*, and *intensity*}

\item{color}{grouping variable used to color traces or curves;
defaults to the *well* variable
(if the assay requries the instrument lid to be opened for reagent addition after temperature equilibration);
defaults to *5*}

\item{model}{boolean parameter to choose plotting modeled curves instead of raw data traces;
defaults to FALSE, i.e. plotting raw data}

\item{interactive}{boolean parameter to display a static or interactive plot using `plotly`;
defaults to TRUE}
}
\value{
returns a plot of fluorescence traces or modeled curves, *intensity* vs *cycle*
}
\description{
\code{view_traces} plots either raw fluorescence traces, or modeled curves fit to the raw data.
}
